% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HN.R
\name{getItem}
\alias{getItem}
\title{getItem}
\usage{
getItem(id)
}
\arguments{
\item{id}{integer value corresponding to the item id.}
}
\value{
It return a list of two objects:

   1. a dataframe with tabular data (named "df")

   2. a named list of retrieved kids objects

   3. a named list of retrieved poll-parts objects
}
\description{
\code{getItem} is a function to retrieve data for a specific item.
}
\examples{
# Get a single item
job1 <- getItem(19567427)
story1 <- getItem(19553941)

# Look into the object elements
job1$df
story1$df
story1$kids
}
