% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HN.R
\name{getUsers}
\alias{getUsers}
\title{getUsers}
\usage{
getUsers(users)
}
\arguments{
\item{users}{string or list of user id(s).}
}
\value{
The function returns a list of two objects:

   1. a dataframe with tabular data (named "df")

   2. a named list of submitted items
}
\description{
\code{getUsers} is a function to retrieve data for a specific user or a vector of user IDs.
}
\examples{
# Get a single user data
user1 <- getUsers("jl")
View(user1$df)
user1$submitted

# Get last updated users profiles
updates <- getHN(what = "updates", n = 10)
users <- getUsers(updates$profiles)

View(users$df)
for(id in users$df$id){
   print(users$submitted[id])
}
}
