% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HN.R
\docType{package}
\name{RHackerNews}
\alias{RHackerNews}
\alias{RHackerNews-package}
\title{RHackerNews}
\description{
RHackerNews is a package that provides a basic R client to the Hacker News API.
}
\details{
More info about HackerNews project can be found at at https://news.ycombinator.com/

More info about HackerNews API is available at https://github.com/HackerNews/API

The package collects basic functions to collect data from the main available endpoints as specified
by the v0 version of the API, as of April 2019.

The author is not affiliated with HackerNews or YCombinator.
}
